changed java version. It's now on J2SE-1.5

+++ 2009-newBirtCharts/trunk/plugins/org.eclipse.birt.newcharts/src/org/eclipse/birt/chart/model/newtype/render/DonutRenderer.java
@@ -62,7 +62,7 @@ public class DonutRenderer {
 	private DonutSlice[] sliceList;
 	private DonutSeries donutseries;
 	private int explosion;
-	private int rotation;
+	private double rotation;
 	private double ratio;
 	private double deviceScale;
 	private Palette seriesPalette;
